refactored and split the CommentCard component into smaller components which are the LikesCounter and the ReplyButton. I have also implemented fetching the comments data from the data.json file so that each comment can have its unique CommentCard component.
